+++ packages/eslint-plugin-pf-codemods/index.js
@@ -1,7 +1,11 @@
 const createListOfRules = (version) => {
   const rules = {};
   require("glob")
-    .sync(require.resolve('@patternfly/eslint-plugin-pf-codemods').replace('index.js', `lib/rules/v${version}/*.js`))
+    .sync(
+      require
+        .resolve("@patternfly/eslint-plugin-pf-codemods")
+        .replace("index.js", `lib/rules/v${version}/*.js`)
+    )
     .forEach(function (file) {
       const ruleName = /.*\/([^.]+)/.exec(file)[1];
       rules[ruleName] = require(`./lib/rules/v${version}/${ruleName}`);
@@ -19,7 +23,7 @@ const warningRules = [
   "charts-tooltip-warning",
   "datePicker-warn-appendTo-default-value-changed",
   "horizontalSubnav-ariaLabel",
-  "onToggle-warn-event",
+  "label-warn-truncated-default",
   "nav-warn-flyouts-now-inline",
   "popover-appendTo-default",
   "react-dropzone-warn-upgrade",