+++ packages/eslint-plugin-pf-codemods/lib/rules/v5/dataList-updated-callback.js
@@ -1,0 +1,7 @@
+const { addCallbackParam } = require("../../helpers");
+
+// https://github.com/patternfly/patternfly-react/pull/8723
+module.exports = {
+  meta: { fixable: "code" },
+  create: addCallbackParam(["DataList"], { onSelectDataListItem: "_event" }),
+};