+++ packages/eslint-plugin-pf-codemods/lib/rules/v5/datalist-remove-props.js
@@ -1,0 +1,19 @@
+const {renameProp} = require("../../helpers");
+
+// https://github.com/patternfly/patternfly-react/pull/8388
+
+module.exports = {
+  meta: {fixable: "code"},
+  create: renameProp(
+    "DataList",
+    {
+      onDragFinish: "",
+      onDragStart: "",
+      onDragMove: "",
+      onDragCancel: "",
+      itemOrder: "",
+    },
+    (node, attribute) =>
+    attribute.name?.name.includes("onDrag") ? `${attribute.name.name} prop for ${node.name.name} has been removed. Implement drag and drop using the DragDrop component instead.` :  `${attribute.name.name} prop for ${node.name.name} has been removed.`
+  ),
+};