+++ packages/eslint-plugin-pf-codemods/lib/rules/v5/onToggle-updated-paramaters.js
@@ -1,0 +1,15 @@
+const { addCallbackParam } = require("../../helpers");
+const onToggleAPIUpdateList = [
+  "ApplicationLauncher",
+  "BadgeToggle",
+  "DropdownToggle",
+  "KebabToggle",
+  "Toggle",
+  "Select",
+  "SelectToggle",
+];
+// https://github.com/patternfly/patternfly-react/pull/8667
+module.exports = {
+  meta: { fixable: "code" },
+  create: addCallbackParam(onToggleAPIUpdateList, { onToggle: "_event" }),
+};