+++ packages/eslint-plugin-pf-codemods/test/rules/v5/label-remove-isTruncated.js
@@ -1,0 +1,24 @@
+const ruleTester = require('../../ruletester');
+const rule = require('../../../lib/rules/v5/label-remove-isTruncated');
+
+ruleTester.run("label-remove-isTruncated", rule, {
+  valid: [
+    {
+      code: `import { Label } from '@patternfly/react-core'; <Label />`,
+    },
+    {
+      // No @patternfly/react-core import
+      code: `<Label isTruncated />`,
+    }
+  ],
+  invalid: [
+    {
+      code:   `import { Label } from '@patternfly/react-core'; <Label isTruncated />`,
+      output: `import { Label } from '@patternfly/react-core'; <Label  />`,
+      errors: [{
+        message: `isTruncated prop has been removed for Label. This is now the default behavior. Note that there is also a new property (maxTextWidth) to customize when truncation will occur.`,
+        type: "JSXOpeningElement",
+      }]
+    }
+  ]
+});