+++ packages/eslint-plugin-pf-codemods/test/rules/v5/popper-remove-popperMatchesTriggerWidth.js
@@ -1,0 +1,24 @@
+const ruleTester = require('../../ruletester');
+const rule = require('../../../lib/rules/v5/popper-remove-popperMatchesTriggerWidth');
+
+ruleTester.run("popper-remove-popperMatchesTriggerWidth", rule, {
+  valid: [
+    {
+      code: `import { Popper } from '@patternfly/react-core'; <Popper />`,
+    },
+    {
+      // No @patternfly/react-core import
+      code: `<Popper popperMatchesTriggerWidth />`,
+    }
+  ],
+  invalid: [
+    {
+      code:   `import { Popper } from '@patternfly/react-core'; <Popper popperMatchesTriggerWidth />`,
+      output: `import { Popper } from '@patternfly/react-core'; <Popper  />`,
+      errors: [{
+        message: `popperMatchesTriggerWidth prop has been removed for Popper.  The width can instead be modified via the new minWidth, maxWidth, and width properties`,
+        type: "JSXOpeningElement",
+      }]
+    }
+  ]
+});